full-stack/postgresql-django-react/5-2-react-django-authentication/django/requirements.txt
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.1
django-cors-headers==4.7.0
django-filter==24.3
djangorestframework==3.15.2
google-auth==2.38.0
idna==3.10
Markdown==3.7
pyasn1==0.6.1
pyasn1_modules==0.4.1
requests==2.32.3
rsa==4.9
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.3.0
